requirements.txt
discord.py==1.2.4
requests==2.22.0
requests_cache==0.5.2
fulltmdb
